Scheduling Jobs
- Add a new task in ReminderTasks for issuing the charging reminder
- Define a task using TaskManager.DefineTask that runs the charging reminder task off the main thread
- Schedule the task by its task tag using the Background.registerTaskAsync
- remove the Button used to test notification that just trigger the notification right away. Now its scheduled

Few FYIs
- The way we scheduled the job is not as ideal but it works. Basically, when the user installs the up, we schedule a task that runs every 15 minutes, and check if the phone is charging, and if its charging, it send the user a reminder to drink water.
If you look at typical android jobDispatcher, you can choose to only dispatch the job you scheduled only when the phone is charging. There are other package in the react directory that help you do that but its not on iOS. So there is no need to stress much on those. I decided to just stick withExpo as i was thought!
https://github.com/vikeri/react-native-background-job
https://rossbulat.medium.com/react-native-background-task-management-in-ios-d0f05ae53cc5