Show when charging : Dynamic Boradcast receivers
- create a showCharging method which switches the icon between gray and pink
- Add an evenLIstener Battery.addBatteryStateListener that listens for if the device is charging or not charging.
- create a state called batteryState that is used to render updates in the UI to pink if it was triggered because the phone was charging

We used a dynamic broadcast receiver to change the UI depending on whether the device is plugged in or not when the app is not plugged in or not.
When the device is plugged in, we show the pink icon and when its not plugged in we show the gray icon

Dynamic broadcast receives is what is advised to be used for your app rather than static broadcast receivers. A static broadcast receiver is where you basically add the broadcast reciver in your AndroidManifest file. It is very rare where you do that

Eg of other broadcast revivers are
- Network
- HEADSET_PLUG
- SCREEN_ON
- DOWNLOAD_COMPLETE
etc...

You can listen for these broadcasts whether your device is running or not running like we demonstrated in this project!

Feel free to browse through expo SDK to see what broadcast receivers that available for you to use!
https://docs.expo.io/versions/v41.0.0/sdk/netinfo/